updated - latest code
updated code

+++ apps/students/models.py
@@ -77,7 +77,7 @@ class Student(models.Model):
 
 
     
-    passport = models.ImageField("Photo",blank=True, upload_to="students/passports/")
+    passport = models.ImageField("Photo",blank=False, upload_to="students/passports/")
     address = models.CharField("Address", max_length=255,default=None,blank=True)
     address1 = models.CharField("Address Line 2", max_length=255,default=None,blank=True,null=True)
     address2 = models.CharField("Address Line 3", max_length=255,default=None,blank=True,null=True)
@@ -109,16 +109,24 @@ class Student(models.Model):
 
     def get_absolute_url(self):
         return reverse("student-detail", kwargs={"pk": self.pk})
+    
 
-    def save(self, *args, **kwargs):
-        if not self.user:
-            self.user = User.objects.create_user(username=self.username,password=self.password,is_staff=False)
-        super().save(*args, **kwargs)
+
+    def save(self, *args,from_save_update = True, **kwargs):
+        if from_save_update:
+            
+            if not self.user:
+                self.user = User.objects.create_user(username=self.username,password=self.password,is_staff=False)
+            super().save(*args, **kwargs)
+        elif from_save_update == False:
+            self.user.delete()
+            self.user = None
+            super().save(*args, **kwargs)
     
     
     def delete(self, *args, **kwargs):
         # Delete the associated user before deleting the Staff instance
-        self.user.delete()
+        self.save(from_save_update=False)
         super().delete(*args, **kwargs)
